+++ src/guis/zeale/guis/developer_page/DeveloperApp.java
@@ -12,9 +12,9 @@ import javafx.scene.input.ScrollEvent;
 import javafx.scene.input.ScrollEvent.HorizontalTextScrollUnits;
 import javafx.scene.input.ScrollEvent.VerticalTextScrollUnits;
 import javafx.scene.layout.Pane;
+import krow.fx.scene.HorizontalMultiScrollBox;
+import krow.fx.scene.HorizontalScrollBox;
 import krow.guis.GUIHelper;
-import krow.scene.HorizontalMultiScrollBox;
-import krow.scene.HorizontalScrollBox;
 import kröw.core.Kröw;
 import kröw.gui.Application;
 import kröw.gui.ApplicationManager;
